+++ location_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1] - 24th August 2020
+
+- Add grantedLimited enum value to PermissionStatus to capture the new limited range authorization that iOS 14 is offering.
+
 ## [1.0.0] - 26th March 2020
 
 - Created the platform interface of the Location plugin in order to support Web and macOS (huge thanks to long1eu)